+++ app/containers/references/Details.js
