+++ app/containers/references/Details.scss
