+++ app/reducers/index.js
@@ -1,11 +1,9 @@
 import { combineReducers } from 'redux';
 import { routerReducer as routing } from 'react-router-redux';
-import counter from './counter';
-import itemDetails from './itemDetails';
+import tagList from 'reducers/tags';
 
 const rootReducer = combineReducers({
-  counter,
-  itemDetails,
+  tagList,
   routing,
 });
