+++ app/reducers/references/itemDetails.js
