+++ webpack.config.base.js
@@ -18,7 +18,10 @@ export default {
   },
   resolve: {
     extensions: ['', '.js', '.jsx', '.json'],
-    packageMains: ['webpack', 'browser', 'web', 'browserify', ['jam', 'main'], 'main']
+    packageMains: ['webpack', 'browser', 'web', 'browserify', ['jam', 'main'], 'main'],
+    root: [
+      path.resolve('./app/'),
+    ]
   },
   plugins: [
